backend/requirements.txt
fastapi
uvicorn
av
numpy
openai-whisper
soundfile
pyinstaller
flask
flask_cors